moved gps satellite listener as per review
