+++ android/src/main/java/org/mozilla/mozstumbler/client/mapview/maplocation/UserPositionUpdateManager.java
@@ -5,52 +5,22 @@
 package org.mozilla.mozstumbler.client.mapview.maplocation;
 
 import android.content.Context;
-import android.location.GpsSatellite;
-import android.location.GpsStatus;
 import android.location.Location;
 import android.location.LocationManager;
 import android.widget.Toast;
-import java.lang.ref.WeakReference;
 import org.mozilla.mozstumbler.client.MainApp;
 import org.mozilla.mozstumbler.client.mapview.MapFragment;
 import org.mozilla.mozstumbler.service.AppGlobals;
 import org.mozilla.mozstumbler.service.core.logging.ClientLog;
-import org.mozilla.mozstumbler.service.stumblerthread.scanners.GPSScanner;
 import org.mozilla.mozstumbler.svclocator.services.log.LoggerUtil;
 
 
 public class UserPositionUpdateManager {
     private static final String LOG_TAG = LoggerUtil.makeLogTag(UserPositionUpdateManager.class);
-    private final WeakReference<MapFragment> mMapFragment;
     private LocationManager mLocationManager;
     private final MultiSourceLocationListener mMultiSourceLocationListener;
 
-    private final GpsStatus.Listener mSatelliteListener = new GpsStatus.Listener() {
-        public void onGpsStatusChanged(int event) {
-            if (event == GpsStatus.GPS_EVENT_SATELLITE_STATUS && mLocationManager != null) {
-                GpsStatus gpsStatus = mLocationManager.getGpsStatus(null);
-                Iterable<GpsSatellite> sats = gpsStatus.getSatellites();
-                int satellites = 0;
-                int fixes = 0;
-                for (GpsSatellite sat : sats) {
-                    satellites++;
-                    if (sat.usedInFix()) {
-                        fixes++;
-                    }
-                }
-                if (mMapFragment.get() != null) {
-                    mMapFragment.get().updateGPSInfo(satellites, fixes);
-                }
-
-                if (fixes < GPSScanner.MIN_SAT_USED_IN_FIX) {
-                    mMultiSourceLocationListener.notEnoughSatellites();
-                }
-            }
-        }
-    };
-
     public UserPositionUpdateManager(MapFragment mapFragment) {
-        mMapFragment = new WeakReference<MapFragment>(mapFragment);
         Context context = mapFragment.getActivity().getApplicationContext();
         mLocationManager = (LocationManager) context.getSystemService(Context.LOCATION_SERVICE);
         mMultiSourceLocationListener = new MultiSourceLocationListener(mLocationManager, mapFragment);
@@ -61,15 +31,13 @@ public class UserPositionUpdateManager {
             return;
         }
 
-        mLocationManager.addGpsStatusListener(mSatelliteListener);
-
         Location lastLoc = mLocationManager.getLastKnownLocation(LocationManager.GPS_PROVIDER);
         if (lastLoc == null) {
             lastLoc = mLocationManager.getLastKnownLocation(LocationManager.NETWORK_PROVIDER);
         }
-        if (lastLoc != null && mMapFragment.get() != null) {
+        if (lastLoc != null) {
             ClientLog.d(LOG_TAG, "set last known location");
-            mMapFragment.get().setUserPositionAt(lastLoc);
+            mapFragment.setUserPositionAt(lastLoc);
         }
 
         MainApp app = mapFragment.getApplication();
@@ -101,8 +69,6 @@ public class UserPositionUpdateManager {
         if (mLocationManager == null) {
             return;
         }
-
-        mLocationManager.removeGpsStatusListener(mSatelliteListener);
         mMultiSourceLocationListener.stop();
     }
 }
